move Motion and Operation into Accelerator
